Refactor category and product types into separate modules; add CategoryPage route guard for invalid categories; update HomePage to compute brand new and hot prices locally; clean up API and hooks exports
