+++ src/api/index.ts
@@ -1,4 +1,4 @@
-import { Product, ProductDetails } from '../types';
+import { Product, ProductDetails, Category } from '../types';
 
 const BASE_URL = '/api';
 
@@ -24,17 +24,7 @@ export const getTablets = (): Promise<ProductDetails[]> =>
 export const getAccessories = (): Promise<ProductDetails[]> =>
   get<ProductDetails[]>('accessories.json');
 
-export const getProductsByCategory = (
-  category: 'phones' | 'tablets' | 'accessories',
-): Promise<Product[]> =>
-  getProducts().then(products => products.filter(p => p.category === category));
-
-export const getBrandNew = (): Promise<Product[]> =>
-  getProducts().then(products => [...products].sort((a, b) => b.year - a.year));
-
-export const getHotPrices = (): Promise<Product[]> =>
+export const getProductsByCategory = (category: Category): Promise<Product[]> =>
   getProducts().then(products =>
-    [...products]
-      .filter(p => p.fullPrice > p.price)
-      .sort((a, b) => b.fullPrice - b.price - (a.fullPrice - a.price)),
+    products.filter(product => product.category === category),
   );
