+++ src/hooks/index.ts
@@ -1,1 +1,3 @@
-// Custom hooks will be defined here
+export * from './useQueryParams';
+export * from './useInterval';
+export * from './useProductCounts';
